+++ CHANGELOG.md
@@ -7,9 +7,11 @@
 
 ### Added
 - Добавлено отображение иконок `.exe` вместо обложек на компактной детальной странице
+- Добавлено редактирование ярлыков Steam по AppID
 
 ### Changed
 - Кэш EXE-иконок унифицирован для карточек и детальных страниц
+- Папки `custom_data` теперь используют имя exe и MD5 полного пути для предотвращения конфликтов одинаковых имён
 - Обновлён перевод на Немецкий `https://translate.codeberg.org/projects/portprotonqt/portprotonqt/de/`
 - Обновлён перевод на Португальский (Бразилия) `https://translate.codeberg.org/projects/portprotonqt/portprotonqt/pt_BR/`
 - Обновлён перевод на Китайский (упрощённый) `https://translate.codeberg.org/projects/portprotonqt/portprotonqt/zh_Hans/`